[DCP][BE] Move DCP._state_dict_utils out from DCP
DCP._state_dict_utils is also used by FSDP. This can cause circular import sometimes. Move it out from DCP to avoid circular import.

Differential Revision: [D52022440](https://our.internmc.facebook.com/intern/diff/D52022440/)

[ghstack-poisoned]
