+++ test/distributed/_composable/fully_shard/test_fully_shard_model_checkpoint.py
@@ -9,7 +9,7 @@ import torch
 import torch.distributed as dist
 import torch.nn as nn
 from torch.distributed._composable import fully_shard
-from torch.distributed.checkpoint._state_dict_utils import _gather_state_dict
+from torch.distributed._state_dict_utils import _gather_state_dict
 from torch.distributed.fsdp import FullyShardedDataParallel as FSDP, StateDictType
 from torch.distributed.fsdp.api import ShardingStrategy
 from torch.distributed.fsdp.wrap import ModuleWrapPolicy
