+++ test/distributed/checkpoint/test_state_dict_utils.py
@@ -6,7 +6,7 @@ import torch.distributed._functional_collectives as funcol
 
 from torch.distributed._tensor import DTensor
 from torch.distributed._tensor.placement_types import Shard
-from torch.distributed.checkpoint._state_dict_utils import (
+from torch.distributed._state_dict_utils import (
     _gather_state_dict,
     _offload_state_dict_to_cpu,
 )
