+++ torch/testing/_internal/distributed/common_state_dict.py
@@ -9,7 +9,7 @@ import torch.nn as nn
 
 from torch.distributed._sharded_tensor import ShardedTensor
 from torch.distributed._tensor import DTensor
-from torch.distributed.checkpoint._state_dict_utils import _gather_state_dict
+from torch.distributed._state_dict_utils import _gather_state_dict
 from torch.distributed.checkpoint.state_dict import (
     PG,
     set_state_dict,
